pyproject.toml
[tool.poetry]
name = "delight"
version = "0.1.0"
description = ""
authors = ["kkkirill"]

[tool.poetry.dependencies]
python = "^3.8"
botocore = "==1.19.18"
boto3 = "==1.16.18"
Pillow = "==8.0.1"
Django = "==3.1.3"
djangorestframework = "==3.12.2"
psycopg2-binary = ">=2.7,<3.0"
django-filter = "==2.4.0"
factory-boy = "==3.1.0"
django-rest-auth = "==0.9.5"
django-redis = "==4.12.1"
celery = "==5.0.2"
drf-yasg = "==1.20.0"
packaging = "==20.4"
django-storages = "==1.10.1"
python-dotenv = "==0.15.0"
django-cors-headers = "==3.5.0"
djangorestframework-camel-case = "==1.2.0"
drf-extensions = "==0.7.0"
elasticsearch = ">=7.0.0,<8.0.0"
django-elasticsearch-dsl = "==7.2.0"
django-elasticsearch-dsl-drf = "==0.22"

[tool.poetry.dev-dependencies]
pytest = "==6.1.2"
pytest-factoryboy = "==2.0.3"
pytest-django = "==4.1.0"
localstack = "==0.12.2"
awscli = "==1.18.178"
boto3 = "==1.16.18"
pylama = "^7.7"
isort = "^4.3"
black = "20.8b1"

[tool.black]
exclude = '.vscode|.eggs|.venv'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
